assistance/pdf/retriever/requirements.txt
Flask
Flask-CORS
langchain
langchain-community
langchain-qdrant
langchain-text-splitters
qdrant-client
sentence-transformers
groq
python-dotenv
gunicorn